requirements.txt
aiogram==2.25.2
python-dotenv==1.1.0
